Add is_ingesting flag to system stats
