+++ crates/libs/common_services/src/api/jobs/user_level.rs
@@ -183,24 +183,3 @@ pub async fn retry_user_job(
 
     Ok(())
 }
-
-/// Checks whether the user has active ingest tasks scheduled or actively running.
-/// -- removed `ingest_llm` from this list
-pub async fn is_user_ingesting(pool: &PgPool, user_id: i32) -> Result<bool, AppError> {
-    let exists = sqlx::query_scalar!(
-        r#"
-        SELECT EXISTS (
-            SELECT 1
-            FROM jobs
-            WHERE user_id = $1
-              AND status IN ('queued'::job_status, 'running'::job_status)
-              AND job_type IN ('ingest_metadata', 'ingest_thumbnails', 'ingest_analysis')
-        ) as "exists!"
-        "#,
-        user_id
-    )
-        .fetch_one(pool)
-        .await?;
-
-    Ok(exists)
-}
