+++ crates/libs/common_services/src/api/system/service.rs
@@ -43,10 +43,13 @@ pub async fn get_system_stats(
         })
     });
 
-    let (db_res, fs_res) =
-        tokio::try_join!(async { db_task.await.map_err(AppError::from) }, async {
-            fs_task.await.map_err(AppError::from)?
-        })?;
+    let ingesting_task = is_user_ingesting(pool, user_id);
+
+    let (db_res, ingest_res, fs_res) = tokio::try_join!(
+        async { db_task.await.map_err(AppError::from) },
+        ingesting_task,
+        async { fs_task.await.map_err(AppError::from)? }
+    )?;
 
     Ok(SystemStats {
         has_clustered_people: db_res.has_people,
@@ -54,5 +57,27 @@ pub async fn get_system_stats(
         allow_file_deletion: constants().allow_file_deletion,
         allow_file_modifications: constants().allow_file_modifications,
         disk: fs_res,
+        is_ingesting: ingest_res,
     })
+}
+
+/// Checks whether the user has active ingest tasks scheduled or actively running.
+/// -- removed `ingest_llm` from this list
+pub async fn is_user_ingesting(pool: &PgPool, user_id: i32) -> Result<bool, AppError> {
+    let exists = sqlx::query_scalar!(
+        r#"
+        SELECT EXISTS (
+            SELECT 1
+            FROM jobs
+            WHERE user_id = $1
+              AND status IN ('queued'::job_status, 'running'::job_status)
+              AND job_type IN ('ingest_metadata', 'ingest_thumbnails', 'ingest_analysis')
+        ) as "exists!"
+        "#,
+        user_id
+    )
+    .fetch_one(pool)
+    .await?;
+
+    Ok(exists)
 }
